+++ include/AutopilotManagerConfig.hpp
@@ -16,9 +16,9 @@ public:
 	bool WriteToFile(const std::string &config_path) const;
 	bool InitFromFile(const std::string &config_path);
 
-	uint32_t autopilot_manager_enabled = false;
+	uint8_t autopilot_manager_enabled = false;
 	std::string decision_maker_input_type = "";
-	uint32_t simple_collision_avoid_enabled = false;
+	uint8_t simple_collision_avoid_enabled = false;
 	double simple_collision_avoid_distance_threshold = 0.0;
 	std::string simple_collision_avoid_distance_on_condition_true = "";
 	std::string simple_collision_avoid_distance_on_condition_false = "";
